requirements.txt
# Telegram API and scraping
telethon==1.32.1
python-telegram-bot==20.7

# Data processing and manipulation
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0

# Text processing and NLP
transformers==4.36.2
tokenizers==0.15.0
nltk==3.8.1

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Database and storage
json5==0.9.14

# Utilities
tqdm==4.66.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Jupyter notebooks
jupyter==1.0.0
ipykernel==6.27.1

# Development and testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0
